delete old scoring component and replace it with new one
