+++ frontend/src/app/keyresult/keyresult.component.html
@@ -1,6 +1,5 @@
 <div (click)="openDrawer()" class="key-result rounded-2 p-3">
   <div class="lh-base">{{ keyResult.title }}</div>
-  <app-scoring [keyResult]="keyResult"></app-scoring>
   <app-scoring2 [keyResult]="keyResult"></app-scoring2>
   <div class="d-flex justify-content-between">
     <div class="d-flex justify-content-center">
