--- frontend/src/app/shared/scoring/scoring/scoring2.component.scss
@@ -1,46 +1,0 @@
-.okr-score-lable {
-  display: inline-grid;
-  width: 30%;
-}
-
-.okr-score {
-  display: inline-grid;
-  height: 21px;
-  width: 30%;
-  background-color: white;
-  outline: 1px solid #5d6974;
-}
-
-.okr-score-commit {
-  width: 40%;
-}
-
-.okr-stretch {
-  margin-left: 16px;
-}
-
-.starIcon {
-  display: block;
-  width: 21px;
-  height: 21px;
-}
-
-.score-red {
-  background-color: #ba3838;
-}
-
-.score-yellow {
-  background-color: #ffd600;
-}
-
-.score-green {
-  background-color: #1e8a29;
-}
-
-.border-right {
-  border-right: 1px solid gray;
-}
-
-.score-stretch {
-  background-image: url("../../../../assets/images/scoring-stars.svg");
-}
